Changes notification ordering for new interruption model.
This change modifies NotificationComparator such that when the new
interruption model is enabled, notifications are sorted with the
primary criterion being whether they are importance >= DEFAULT,
outweighing other criteria that are currently considered first
(colorization, ongoingness, messaging, people).  This allows us
to ensure that high priority (>= DEFAULT) and low priority
(< DEFAULT) notifications are contiguous so that thedisplay of them
as two distinct groups makes sense.

Test: atest NotificationComparatorTest
Change-Id: Ifa1405546b75b9da4b8411f861675cd47793785c