Add SettingsPanel Intent Actions
A Settings Panel is a dialog that contains a fixed subset of
settings to address a particular user problem. The ultimate
goal is that users can solve problems in-app rather than going
into settings. This is limitted to to "safe" settings which do
not need the full context of the settings.

The Settings are hosted in a Settings activity, which prevents
scraping from the calling app.

Test: make -j40 RunSettingsRobotests, manual app
Test: atest SettingsPanelTest
Fixes: 117804442
Change-Id: I3afb9e235959d0e4bc889747c4518de89918674c