Checking a job's source user id to determine readiness.
We should be checking the source user id to determine if a job is ready
to run instead of the calling user id. The calling user id is primarily
kept to validate who can schedule and cancel a job, but a job should
only run if the source user is started.

Bug: N/A
Test: atest CtsJobSchedulerTestCases PrioritySchedulingTest
Change-Id: Idc5ee44c2ce3c5cb5300f38e5eec81709afed3b8