Do not report resized before sending new configuration
When display size changed, we reported to client that its windows
should be resized (but with old configuration). The new configuration
was then updated later. Since the IWindow#resized is a one way
binder call, there is no guarantee that client has completed
scheduling window-resized messages before the new configuration
updating request. So, test case was flaky while running on CF, which
had failed if the client updated to new configuration prior to handle
window resized.

Bug: 121105810
Test: atest ActivityManagerMultiDisplayTests

Change-Id: Ic661bc36494d268a2e712a079ba26baf47626130