Don't make active activities that are launched behind
Activities that are being launched behind are not intended to be
presented to the user, so they should not be made active (resumed
or paused) based just on their visibility and position.

One example is home/recents activity during transition from an
app to home after home button press with QuickStep enabled - the
home stack is made visible for the duration of the transition,
but it's still behind a fullscreen app stack and should not be
resumed immediately.

Bug: 123166999
Bug: 117135575
Test: ActivityManagerActivityVisibilityTests#testGoingHomeMultipleTimes
Test: ActivityManagerActivityVisibilityTests#testPressingHomeButtonMultipleTimes
Test: ActivityManagerActivityVisibilityTests#testPressingHomeButtonMultipleTimesQuick
Test: Manually switch between home and app with QuickStep enabled
Change-Id: I1106cf3da923e60d02b84d5fe4a552865113afc8