Make hidden constructor of PlaybackInfo as @SystemApi
When the android build setting is configured to a rule which moves
MediaSession1 related codes to mainline module, the non-public and
non-system APIs in the module cannot be used by the framework.

This CL makes the hidden constructor of PlaybackInfo as
system API, so that it can be used by the framework.

Bug: 123500940
Test: make update-api -j; make -j;
Change-Id: I8583d6f79c70f8bbbbedfa0a75676218aeec58e6