Revert "Fix FDE"
This reverts commit b1f313b535c04dc1fdbefd0eda7a7ebccc575dba.

Wifi service is assumed to be always up by other parts of the Android
system. So, not starting wifi service will cause NPE crashes all over
the place.

Wifi service already goes into idle mode to handle this FDE transition,
so there is no need for it to be turned off.

Bug: 78663946
Test: Manually verified non-FDE to FDE transition on Taimen (Using
ag/6180881).

Change-Id: I2d4d38e978a671f4dcd671524a1b38f557510bf5