[Magnifier-76] Show for translated/scaled textview
In Ibb81fdc9d2ec8ba14914166e408c92a3aad7e312 we disabled the magnifier
for cases when the magnified TextView (or a view above it in the view
hierarchy) was matrix transformed (translated, scaled or rotated).
This CL reenables the magnifier for translated and scaled TextViews. The
translation case was already working.

Bug: 123536409
Bug: 77283509
Test: manual testing
Change-Id: I1a22ff2223fe699c6ac660b5f59510ca1f544ee4