Add callback to sendCecCommand in SendKeyEventAction when handling volume key forwarding.
The device that is not directly connected with audio system device
is not able to detect when the audio system is unplugged. So the send
key event action needs to callback when the message can't be sent to
audio system any more (which means it's unglugged). The action should
also update the sam status because of the absence of audio system
device.

Test: manual
Bug: 123369653
Change-Id: I6259bad26b96622aebcf3430e770211c32a92f1b
(cherry picked from commit f1e61c1895e3a87e44e71a9c2f29f0077af58df0)